Reorient namespaces to include 'early.' prefix.
This sort of thing should not be done often, or indeed hardly ever,
once we reach a higher level of maturity -- it's *not* following the
"append only" mutation rules.  But I want to get this prefix in now,
early (heh), to indicate the overall "v0"-ness of these data.

We'll still plan to carry this data a long time.  But it's possible
that we'll later port subsets of it to non-"early." namespaces, and
take that opportunity to review naming conventions, etc, which might
accumulate some new guidelines after we've been at this a while and
gained some practical experience with what works smoothly.

Signed-off-by: Eric Myhre <hash@exultant.us>