Switch version name of debian image to v1.* prefix
I don't know any better way to name those than with their date; but it
seems wise to put a leading number that's *not* in the thousands, in
case we ever want to switch to a different versioning scheme later and
have it start with numbers below 4000.

Like the previous commit: this is the sort of thing that should not be
done whimsically, or at all, in the future when we begin to take the
"append only" rules more seriously.

Signed-off-by: Eric Myhre <hash@exultant.us>