Scans of a variety of upstream sources.
These are all things I tried to have some fun with this weekend.

The "gcc-plus" monolith image is from a previous round of work in the
now-archived "refreshing" repo, and we probably *will not* be able to
reproduce that snapshot, as if I recall correctly, the GCC version
selected therein was from latest(at-the-time) rather than snapshot.d.o.
This is of course dissatisfying, but it's a part of history that
remains currently useful.

The other things appearing here in the "hyphae" group are clear ingests
from their upstreams.  The versions of interest are the ones from the
time of the debian/debuerreotype captures in the now-archived
"refreshing" repo -- we're going to try to reproduce the main image,
even if we aren't optimistic about the "gcc-plus" one.

Signed-off-by: Eric Myhre <hash@exultant.us>