Release a "minimal image" which contains ld.so.
And some device nodes and I couldn't bring myself to remove /home
nor /tmp.

The purpose of this is to have a container where we can really properly
test that things have bundled all of their libraries.  Even the
busybash image turns out to have far too many libraries laying around
to be a useful test for this.

And in fact, even the musl ld.so is too powerful for this.  It exports
about half the symbols that you usually expect from libc... from the
interpreter itself.  This kind of blows my mind, but yep, it's a thing.

So, this is the ld.so 2.24, as taken from our other debian base image
monolith.

No libc is in the image.  No nothing.  Just the elf interpreter itself.

Of course, an even higher bar of "portability" would demand that
binaries ought not rely on *anything* at all.  However, this container
is for when you're following the logic tree that follows from conceding
that the ELF interpreter exists, in practice everyone has agreed that
it lives at "/lib64/ld-linux-x86-64.so.2", and that path is simply
going to be with us for a (long, long) while.

Signed-off-by: Eric Myhre <hash@exultant.us>