Rename 'catalog.tl'->'lineage.tl'.
This is a change to naming that should reduce a lot of ambiguity:
a "Lineage" is the list of Releases for a module;
a "Catalog" is the whole tree of collected Lineages.

The reach (nee stellar) tool will now look for these 'lineage.tl'
filenames instead of the 'catalog.tl' files as of the v0.300 release.

If you need to perform this kind of rename on your own files:

find -name catalog.tl | xargs -I {} bash -c 'mv {} $(echo {} | sed s@catalog.tl@lineage.tl@)'

Signed-off-by: Eric Myhre <hash@exultant.us>